Move message utility code to avoid circular package reference
